chore: refactor SpanAggregator <> TraceWriter interface Help remove some implementation details from SpanAggregator into a the writer module.
