+++ ddtrace/internal/writer/__init__.py
@@ -1,3 +1,4 @@
+from .writer import create_trace_writer  # noqa: I001,F401
 from .writer import AgentResponse  # noqa: I001,F401
 from .writer import AgentWriter  # noqa:I001,F401
 from .writer import DEFAULT_SMA_WINDOW  # noqa:F401
